Update deprecated php type casts
In PHP 8.5, non-canonical scalar type casts (boolean|double|integer|binary) are deprecated.

See https://php.watch/versions/8.5/boolean-double-integer-binary-casts-deprecated

+++ CRM/Advancedfundraising/Form/Report/OpenFlashChart.php
@@ -276,7 +276,7 @@ class barchart extends chart {
  */
   function setChartValues(){
     foreach ($this->values as $xVal => $yVal) {
-      $this->yValues[] = (double) $yVal[0];
+      $this->yValues[] = (float) $yVal[0];
       $this->xValues[] = (string) $xVal;
     }
     $this->chartElement->set_values($this->yValues);
